% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateGenes.R
\name{annotateGenes}
\alias{annotateGenes}
\title{This function annotate ENSEMBLE gene IDs into HGNC symbol using the biomaRt
package.}
\usage{
annotateGenes(ensemblID)
}
\arguments{
\item{ensemblID}{ENSEMBL internal gene ID}
}
\value{
data frame with two columns ENSEMBL id and mapped HGNC symbol
}
\description{
This function annotate ENSEMBLE gene IDs into HGNC symbol using the biomaRt
package.
}
\examples{
\dontrun{annotateGenes(ensemblID = "ENSG00000000003")}
}
