% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadCNVdata.R
\name{downloadCNVdata}
\alias{downloadCNVdata}
\title{This function downloads, pre-processes and filters harmonized copy number
variation data obtained via RNA-Seq for a given cohort of patients available
within The Cancer Genome Atlas project.}
\usage{
downloadCNVdata(projectID, data.dir = "./dataDownload")
}
\arguments{
\item{projectID}{cohort of patients in TCGA database,
for list available projects type: listTCGAcohorts(),
character string}

\item{data.dir}{directory to save expression data,
character string, default: "./dataDownload"}
}
\value{
matrix with CNV data;
        columns are internal patient IDs, rows are HGNC gene symbols
}
\description{
This function downloads, pre-processes and filters harmonized copy number
variation data obtained via RNA-Seq for a given cohort of patients available
within The Cancer Genome Atlas project.
}
\examples{
\dontrun{downloadCNVdata(projectID = "TCGA-COAD")}
}
