% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadExpressionData.R
\name{downloadExpData}
\alias{downloadExpData}
\title{This function downloads, pre-processes and filters harmonized expression data
obtained via RNA-Seq for a given cohort of patients available within
The Cancer Genome Atlas project.}
\usage{
downloadExpData(
  projectID,
  data.dir = "./dataDownload",
  PreProc_cor.cut = 0.6,
  Filt_method = "quantile",
  Filt_qnt.cut = 0.25
)
}
\arguments{
\item{projectID}{cohort of patients in TCGA database,
for list available projects type: listTCGAcohorts(),
character string}

\item{data.dir}{directory to save expression data,
character string, default: "./dataDownload"}

\item{PreProc_cor.cut}{threshold to filter samples according their spearman
correlation in samples by samples,
numeric in the range 0-1,
default: 0.6}

\item{Filt_method}{method of filtering,
available: 'quantile', 'varFilter'}

\item{Filt_qnt.cut}{threshold selected as mean for filtering,
numeric in the range 0-1, default: 0.25}
}
\value{
matrix with expression data;
        columns are internal patient IDs, rows are HGNC gene symbols
}
\description{
This function downloads, pre-processes and filters harmonized expression data
obtained via RNA-Seq for a given cohort of patients available within
The Cancer Genome Atlas project.
}
\examples{
\dontrun{downloadExpData(projectID = "TCGA-COAD")}
}
