% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeGenotypeComparison.R
\name{empty_plot}
\alias{empty_plot}
\title{Helper function for visGenotypeComparison.}
\usage{
empty_plot(...)
}
\arguments{
\item{...}{empty}
}
\value{
empty plot
}
\description{
Helper function for visGenotypeComparison.
}
\examples{
\dontrun{empty_plot()}
}
