% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputSLanalyzeR.R
\docType{data}
\name{inputSLanalyzeR}
\alias{inputSLanalyzeR}
\title{Expression data}
\format{
An object of class S3 data frame
}
\usage{
data(inputSLanalyzeR)
}
\description{
Data from The Cancer Genome Atlas project for READ cohort of patients.
161 patients, copy number data for TP53, expression data for TSPAN6
}
\examples{
data(inputSLanalyzeR)
}
\references{
data frame with patient IDs, CNV and expression data
}
\keyword{datasets}
