% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listTCGAcohorts.R
\name{listTCGAcohorts}
\alias{listTCGAcohorts}
\title{This function prints available cohorts within The Cancer Genome Atlas
project.}
\usage{
listTCGAcohorts(...)
}
\arguments{
\item{...}{empty}
}
\value{
character vector with TCGA cohorts in the following format: TCGA-XXXX
}
\description{
This function prints available cohorts within The Cancer Genome Atlas
project.
}
\examples{
\dontrun{listTCGAcohorts()}
}
