% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGenotypeComparison.R
\name{matchData}
\alias{matchData}
\title{Helper function for runGenotypeComparison.}
\usage{
matchData(data, genotype.Class = c("Loss", "Gain", "WT"), gene2)
}
\arguments{
\item{data}{output of the prepareData function}

\item{genotype.Class}{genotype class of gene 1,
possible arguments: WT, Gain, Loss,
character string}

\item{gene2}{gene for which normalized expression data were extracted,
character string}
}
\value{
numeric vector of expression data for a given genotype
}
\description{
Helper function for runGenotypeComparison.
}
\examples{
\dontrun{genotype.Class(data, genotype.Class = "WT", gene2 = "CFLAR")}
}
