% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeExpData.R
\name{normalizeExpData}
\alias{normalizeExpData}
\title{This function performs normalization of expression data.}
\usage{
normalizeExpData(expData, method = c("log", "rlog", "vst"))
}
\arguments{
\item{expData}{matrix with gene counts,
rows are genes, columns are subject IDs}

\item{method}{statistical method of data normalization: log = log2(count+1),
rlog = regularized log transformation,
vst = variance stabilizing transformation}
}
\value{
matrix, rows are genes, columns are subject IDs
}
\description{
This function performs normalization of expression data.
}
\examples{
\dontrun{normalizeExpData(expData = expData_READ, method = "rlog")}
}
