% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{This function extracts relevant data for a pair of genes to run statistical
tests and perform results visualization.}
\usage{
prepareData(gene1, gene2, CNVdata, normExpData)
}
\arguments{
\item{gene1}{gene for which copy number data will be extracted,
character string}

\item{gene2}{gene for which normalized expression data will be extracted,
character string}

\item{CNVdata}{matrix with copy number data,
rows are genes, columns are subject IDs}

\item{normExpData}{matrix with gene counts,
rows are genes, columns are subject IDs}
}
\value{
data frame with the following columns:
        1. patientID
        2. copy number data for gene 1
        3. normalized expression data for gene 2
        4. genotype class of gene 1: amplification/gain, deletion/loss,
           wild type/WT
}
\description{
This function extracts relevant data for a pair of genes to run statistical
tests and perform results visualization.
}
\examples{
\dontrun{prepareData(gene1 = "CDK11A", gene2 = "CFLAR",
                     CNVdata = CNVdata, normExpData = normExpData)}
}
