% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGenotypeComparison.R
\name{runGenotypeComparison}
\alias{runGenotypeComparison}
\title{This function runs statistical analysis (wilcoxon test)
which aims at finding differences between expression of gene 2
among copy number variations of gene 1.}
\usage{
runGenotypeComparison(data, gene1, gene2, threshold = 10)
}
\arguments{
\item{data}{output of the prepareData function}

\item{gene1}{gene for which copy number data were extracted,
character string}

\item{gene2}{gene for which normalized expression data were extracted,
character string}

\item{threshold}{minimal number of patients to run test, default = 10}
}
\value{
data frame with comparison of genotypes and associated p-value
}
\description{
This function runs statistical analysis (wilcoxon test)
which aims at finding differences between expression of gene 2
among copy number variations of gene 1.
}
\examples{
\dontrun{runGenotypeComparison(data,
                               gene1 = "CDK11A",
                               gene2 = "CFLAR",
                               threshold = 10)}
}
