% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeGenotypeComparison.R
\name{visGenotypeComparison}
\alias{visGenotypeComparison}
\title{This function runs statistical analysis (wilcoxon test)
which aims at finding differences between expression of gene 2
among copy number variations of gene 1.}
\usage{
visGenotypeComparison(
  data,
  gene1,
  gene2,
  method = c("log", "rlog", "vst"),
  threshold = 10
)
}
\arguments{
\item{data}{output of the prepareData function}

\item{gene1}{gene for which copy number data were extracted,
character string}

\item{gene2}{gene for which normalized expression data were extracted,
character string}

\item{method}{method used for expression data normalization
log = log2(count+1),
rlog = regularized log transformation,
vst = variance stabilizing transformation}

\item{threshold}{minimal number of patients to visualize results}
}
\value{
boxplot
}
\description{
This function runs statistical analysis (wilcoxon test)
which aims at finding differences between expression of gene 2
among copy number variations of gene 1.
}
\examples{
\dontrun{visualizeGenotypeComparison(data,
                                     gene1 = "CDK11A",
                                     gene2 = "CFLAR",
                                     method = "vst",
                                     threshold = 10)}
}
